agent/requirements.txt
paho-mqtt==1.6.1
scikit-learn==1.7.2
pandas==2.2.0
numpy==1.26.4
influxdb-client==1.40.0
joblib==1.3.2
google-genai